requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1

# AI
groq==0.9.0
fastembed==0.3.1
faiss-cpu==1.8.0

# DB
sqlalchemy==2.0.30
aiosqlite==0.20.0

# Document parsing
pypdf==4.2.0
python-docx==1.1.2

# Utilities
httpx==0.27.0
tenacity==8.3.0
loguru==0.7.2
packaging==24.0